주문 기능 드디어 끝

+++ backend/src/main/java/com/scentelier/backend/entity/Ingredient.java
@@ -11,6 +11,7 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -34,7 +35,7 @@ public class Ingredient {
     private String  description;
 
     @Column(nullable = false)//값입력필수
-    private int stock;
+    private BigDecimal stock;
 
     @Column(name = "is_deleted", columnDefinition = "TINYINT(1) DEFAULT 0")
     private boolean isDeleted;
